+++ src/pages/SignupPage.tsx
@@ -3,13 +3,35 @@ import './SignupPage.css';
 import LogoCloud from '../assets/sign/logoCloud.svg';
 import SignForm from '../components/sign/SignForm';
 import { useNavigate } from 'react-router-dom';
+import { useUserStore } from '../store/userStore';
+import { useShallow } from 'zustand/shallow';
+import { postSignup } from '../apis/user';
 
 const SignupPage: React.FC = () => {
+  const navigate = useNavigate();
+
+  const { email, password, clear } = useUserStore(
+    useShallow((state) => ({
+      email: state.email,
+      password: state.password,
+      clear: state.clear,
+    }))
+  );
+  const onHandleSubmit = async () => {
+    try {
+      await postSignup(email, password);
+      clear();
+      navigate('/login');
+    } catch (err) {
+      alert('회원가입 실패');
+      console.log('회원가입 실패 err:', err);
+    }
+  };
   const navigator = useNavigate();
   return (
     <div className='signup-page'>
       <img src={LogoCloud} alt='LogoCloud' />
-      <SignForm buttonText='Sign Up' />
+      <SignForm buttonText='Sign Up' onSubmitFn={onHandleSubmit} />
       <p onClick={() => navigator('/login')}>
         Switch to <span>Login</span>
       </p>
